my-lambda-function/requirements.txt
flywheel-sdk
boto3
pandas
numpy
geopandas
plotly
google-api-python-client
google-auth
fsspec
requests